Added heightMeasurement

+++ WeatherStation_WPF/DataGraphWindow.xaml.cs
@@ -21,7 +21,7 @@ namespace WeatherStation_WPF
     public partial class DataGraphWindow : Window
     {
         ObservableCollection<KeyValuePair<Int32, double>> Temperature   = new ObservableCollection<KeyValuePair<Int32, double>>();
-        ObservableCollection<KeyValuePair<Int32, UInt64>> Pressure      = new ObservableCollection<KeyValuePair<Int32, UInt64>>();
+        ObservableCollection<KeyValuePair<Int32, double>> Pressure      = new ObservableCollection<KeyValuePair<Int32, double>>();
         ObservableCollection<KeyValuePair<Int32, double>> Humidity      = new ObservableCollection<KeyValuePair<Int32, double>>();
         ObservableCollection<KeyValuePair<Int32, UInt64>> Radiance      = new ObservableCollection<KeyValuePair<Int32, UInt64>>();
 
@@ -48,7 +48,7 @@ namespace WeatherStation_WPF
             }
 
             Temperature.Add(new KeyValuePair<Int32, double>(seconds, DataHolder.i_temperature1));
-            Pressure.Add(new KeyValuePair<Int32, UInt64>(seconds, DataHolder.i_pressureCalculated));
+            Pressure.Add(new KeyValuePair<Int32, double>(seconds, DataHolder.i_pressureCalculated));
             Humidity.Add(new KeyValuePair<Int32, double>(seconds, DataHolder.i_humidityAir));
             Radiance.Add(new KeyValuePair<Int32, UInt64>(seconds, DataHolder.i_radianceCalculated));
         }
